Add HSS schema, IDE inlays & UI improvements
Introduce a full HSS schema, parser and compiler integration (HssSchema, HssSyntax, HssParser, HssCompiler) and recoverable parsing with precise error ranges. Add IDE support: lexer, completions, diagnostics, inlay hints and go-to-definition (HssScriptingAnalyzer, UiLanguageCatalog, Hss* UI classes). Implement clickable resource-location inlays and resolution (ResourceLocationTargets, inlay actions). Improve editor widgets and runtime UI: inlay rendering & measurement, editable-field inlay metrics, font-size scaling, better scroll-wheel routing to deepest listener, dock-tab swap/drag slide behavior and related modifiers. Add tests and an icon asset; update changelog.

+++ runtime/src/main/java/ru/hollowhorizon/hollowengine/client/ui/ide/HollowIdeFontSize.kt
@@ -29,8 +29,8 @@ object HollowIdeFontSize {
         }
 
     /** Applies a wheel notch; [scrollY] follows the usual convention of positive meaning scroll up. */
-    fun zoom(scrollY: Double) {
-        if (scrollY == 0.0) return
-        size += if (scrollY > 0) Step else -Step
+    fun zoom(scrollY: Float) {
+        if (scrollY == 0f) return
+        size += if (scrollY > 0f) Step else -Step
     }
 }
